Added all the stuff from before and even added skyboxes for all the colours
added a skybox all colours and scripted them
cleaned up the code in the quiz manager code
what to do next?:
fix the small bug line on the skybox on the custom made colours
Fix the shadow bug on the ball pit
Work entirely on the main scene with the questions and make everything look nice, the panel, the buttons, the materials on the tables and ball.
add more marker colours
fix collision when ball is hitting the marker and it flies away
add a eraser marker too?
fix text when picking up gun